Recursive lookup and masks
Added support for recursive lookup. This will read and parse
the /etc/resolv.conf and use that as a recursive source.

Added support for masks, this can be used to return different
answers based on the remote ip of the client asking the question
(This is also known as split horizon dns)